Fix imaging and add GMOS spectroscopy

+++ bundle/edu.gemini.itc/src/main/java/edu/gemini/itc/operation/ImagingS2NCalculatable.java
@@ -12,7 +12,7 @@ public interface ImagingS2NCalculatable extends Calculatable {
     double getVarReadout();
     double getExposureTime();
 
-    double numberSourceExposures();
+    int numberSourceExposures();
 
     double singleSNRatio();
     double totalSNRatio();
